+++ src/table/components/TableTotals.tsx
@@ -1,12 +1,22 @@
 import React, { memo, useMemo } from 'react';
+
 import { useContextSelector, TableContext } from '../context';
 import { getTotalsCellStyle } from '../utils/styling-utils';
 import { handleTotalKeyDown } from '../utils/handle-key-press';
 import { removeTabAndFocusCell } from '../utils/accessibility-utils';
 import { StyledHeadRow, StyledTotalsCell } from '../styles';
 import { TableTotalsProps } from '../types';
+import CellText from './CellText';
 
-function TableTotals({ rootElement, tableData, theme, layout, keyboard, selectionsAPI }: TableTotalsProps) {
+function TableTotals({
+  rootElement,
+  tableData,
+  theme,
+  layout,
+  keyboard,
+  selectionsAPI,
+  areBasicFeaturesEnabled,
+}: TableTotalsProps) {
   const { columns, paginationNeeded, totalsPosition, rows } = tableData;
   const headRowHeight = useContextSelector(TableContext, (value) => value.headRowHeight);
   const setFocusedCellCoord = useContextSelector(TableContext, (value) => value.setFocusedCellCoord);
@@ -27,13 +37,20 @@ function TableTotals({ rootElement, tableData, theme, layout, keyboard, selectio
             className="sn-table-cell"
             tabIndex={-1}
             onKeyDown={(e: React.KeyboardEvent<HTMLElement>) => {
-              handleTotalKeyDown(e, rootElement, cellCoord, setFocusedCellCoord, selectionsAPI.isModal());
+              handleTotalKeyDown(
+                e,
+                rootElement,
+                cellCoord,
+                setFocusedCellCoord,
+                selectionsAPI.isModal(),
+                areBasicFeaturesEnabled
+              );
             }}
             onMouseDown={() => {
               removeTabAndFocusCell(cellCoord, rootElement, setFocusedCellCoord, keyboard);
             }}
           >
-            {column.totalInfo}
+            <CellText>{column.totalInfo}</CellText>
           </StyledTotalsCell>
         );
       })}