+++ src/table/utils/copy-utils.ts
@@ -1,0 +1,16 @@
+const copyCellValue = async (evt: any) => {
+  let target = evt.target as HTMLElement;
+
+  if (!target.classList.contains('sn-table-cell')) {
+    target = target.closest('.sn-table-cell') as HTMLElement;
+  }
+  const value = target.querySelector('.sn-table-cell-text')?.textContent;
+
+  try {
+    value && (await navigator.clipboard.writeText(String(value)));
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+export default copyCellValue;