+++ src/table/utils/handle-click.ts
@@ -12,9 +12,9 @@ export const handleClickToFocusBody = (
   keyboard: stardust.Keyboard,
   totalsPosition: TotalsPosition
 ) => {
-  const { rawRowIdx, rawColIdx } = cell;
-  const adjustedRowIdx = totalsPosition === 'top' ? rawRowIdx + 2 : rawRowIdx + 1;
-  removeTabAndFocusCell([adjustedRowIdx, rawColIdx], rootElement, setFocusedCellCoord, keyboard);
+  const { pageRowIdx, pageColIdx } = cell;
+  const adjustedRowIdx = totalsPosition === 'top' ? pageRowIdx + 2 : pageRowIdx + 1;
+  removeTabAndFocusCell([adjustedRowIdx, pageColIdx], rootElement, setFocusedCellCoord, keyboard);
 };
 
 export const handleClickToFocusHead = (
@@ -45,8 +45,10 @@ export const getSelectionMouseHandlers = (
     // run handleClickToFocusBody
     onMouseDown?.(evt as React.MouseEvent<HTMLTableCellElement>);
     // only need to check isSelectable here. once you are holding you want to be able to drag outside the current column
-    if (areBasicFeaturesEnabled && cell.isSelectable)
-      selectionDispatch({ type: SelectionActions.SELECT_MULTI_START, payload: { cell } });
+    if (areBasicFeaturesEnabled && cell.isSelectable) {
+      const mouseupOutsideCallback = () => selectionDispatch({ type: SelectionActions.SELECT_MULTI_END });
+      selectionDispatch({ type: SelectionActions.SELECT_MOUSE_DOWN, payload: { cell, mouseupOutsideCallback } });
+    }
   };
 
   const handleMouseOver = (evt: React.MouseEvent) => {
@@ -55,13 +57,14 @@ export const getSelectionMouseHandlers = (
   };
 
   const handleMouseUp = (evt: React.MouseEvent) => {
-    if (areBasicFeaturesEnabled && evt.button === 0) selectionDispatch({ type: SelectionActions.SELECT_MULTI_END });
+    if (evt.button === 0) {
+      if (areBasicFeaturesEnabled) {
+        selectionDispatch({ type: SelectionActions.SELECT_MOUSE_UP, payload: { cell, evt, announce } });
+      } else if (cell.isSelectable) {
+        selectionDispatch({ type: SelectionActions.SELECT, payload: { cell, evt, announce } });
+      }
+    }
   };
 
-  const handleMouseClick = (evt: React.MouseEvent) => {
-    if (cell.isSelectable && evt.button === 0)
-      selectionDispatch({ type: SelectionActions.SELECT, payload: { cell, evt, announce } });
-  };
-
-  return { handleMouseDown, handleMouseOver, handleMouseUp, handleMouseClick };
+  return { handleMouseDown, handleMouseOver, handleMouseUp };
 };